refactor(admin): split ChatRouterService into planner/llm/predicate-plan stages (max-params 2/3)
ChatRouterService had 7 injected deps (config, registry, routeCache, embedder,
collapsePatterns, intentClassifier, metrics?). The cache/collapse/intent
collaborators were already separate services; this splits the remaining stages
so the constructor lands at ≤3:

- PredicatePlanService (registry, embedder) — predicate snapshot + embedding
  hints; thresholds read from env
- ChatRoutePlannerService (predicatePlan, intentClassifier, collapsePatterns) —
  the local pre-pass (temporal/mention/hints/collapse/intent) + collapse-pattern
  teach
- ChatRouterLlmService (config, metrics?) — the single grounded OpenAI call
- ChatRouterService (planner, llm, routeCache) — orchestrates route/routeMiss,
  skip gate, local-override merge, cache, safeDefault

RouteContext moved to chat-route-context.ts; the planner returns it minus the
cache key, which the orchestrator computes. Public route() + re-exports
(ChatRoute type, classifyIntentLocally/extractPredicateHintsLocally/shouldSkipLLM)
unchanged. intentConfidenceFloor read from env.

tsc + typecheck + lint clean; 700 unit + 128 e2e + 9 jobs green.
